backend/python-service/src/requirements.txt
boto3
botocore
requests
flask
werkzeug
psycopg2-binary
paddleocr
pdf2image
huggingface_hub
paddlepaddle
python-dotenv

